+++ ChangeLog.md
@@ -96,6 +96,7 @@ Nazara Development Kit:
 - Fix CollisionComponent3D initialization (teleportation to their real coordinates) which could sometimes mess up the physics scene.
 - ⚠️ Renamed World::Update() to World::Refresh() for more clarity and to differentiate it from World::Update(elapsedTime)
 - World entity ids are now reused from lowest to highest (they were previously reused in reverse order of death)
+- World now has an internal profiler, allowing to measure the refresh and system update time
 
 # 0.4: